requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.0.0
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
gunicorn==20.0.4
